Adding abstraction for test Network and Transport
Requires sero-warp 1.0.1-SNAPSHOT (Not released)
As per MANGO-1975
