+++ .idea/jarRepositories.xml
@@ -7,6 +7,11 @@
       <option name="url" value="https://repo.maven.apache.org/maven2" />
     </remote-repository>
     <remote-repository>
+      <option name="id" value="github-mango4" />
+      <option name="name" value="GitHub packages (Radix IoT private)" />
+      <option name="url" value="https://maven.pkg.github.com/radixiot/maven" />
+    </remote-repository>
+    <remote-repository>
       <option name="id" value="central" />
       <option name="name" value="Maven Central repository" />
       <option name="url" value="https://repo1.maven.org/maven2" />
@@ -25,6 +30,11 @@
       <option name="id" value="ias-snapshots" />
       <option name="name" value="ias-snapshots" />
       <option name="url" value="https://maven.mangoautomation.net/repository/ias-snapshot/" />
+    </remote-repository>
+    <remote-repository>
+      <option name="id" value="github" />
+      <option name="name" value="GitHub packages (Mango public, not used for now)" />
+      <option name="url" value="https://maven.pkg.github.com/mangoautomation/ma-core-public" />
     </remote-repository>
   </component>
 </project>
