+++ pom.xml
@@ -108,7 +108,7 @@
 		<dependency>
 			<groupId>lohbihler</groupId>
 			<artifactId>sero-warp</artifactId>
-			<version>1.0.0</version>
+			<version>1.0.1-SNAPSHOT</version>
 		</dependency>
 
 		<dependency>
